[feat] Images added to HTML
What?:

Breaking Changes?
