+++ src/css/layout.css
@@ -35,11 +35,7 @@ main {
   box-shadow: 4px 4px 8px 0px rgba(0, 0, 0, 0.75);
 }
 
-.page-header {
-  width: 100%;
-}
-
-.page-body {
+.page-header, .page-body {
   width: 100%;
 }
