+++ src/css/search.css
@@ -8,14 +8,17 @@ input {
   background-color: #f1f1f1;
   padding: 10px;
   font-size: 16px;
+  height: 50px;
 }
 input[type=search] {
   background-color: #f1f1f1;
-  width: 100%;
+  width: 79%;
 }
 button[type=submit] {
-  background-color: DodgerBlue;
+  height: 50px;
+  background-color: var(--lightblue);
   color: #fff;
+  width: 19%;
 }
 .autocomplete-items {
   position: relative;
