+++ src/index.html
@@ -17,28 +17,36 @@
         <main>
             <div class="page-header">
                 <div id="autocomplete" >
-                  <input type="search" name="search" id="search-input" placeholder="City Name">
+                  <input type="search" name="search" id="search-input" placeholder="City Name" autocomplete="address-level2">
                   <button type="submit" id="search-button">Search</button>
                 </div>
             </div>
             <div class="page-body">
                 <div class="grid-container">
                     <div class="humidity">
-                      <div class="humidityIcon">icon</div>
+                      <div class="humidityIcon">
+                        <img src="src/img/humidity.png" alt="humidity">
+                      </div>
                       <div class="humidityNumber">50</div>
                     </div>
                     <div class="date">
                       <div class="dayName">Monday</div>
+                      <div class="dayNumber">07/03/2020</div>
                     </div>
                     <div class="place">
                       <div class="city">Chinu</div>
+                      <div class="country">CO</div>
                     </div>
-                    <div class="temperatures">80</div>
-                    <div class="cloudiness">56</div>
+                    <div class="temperatures">
+                      <img src="src/img/hot.png" alt="hot">39
+                    </div>
+                    <div class="cloudiness">
+                      <img src="src/img/cloud.png" alt="cloud">56%
+                    </div>
                     <div class="moreInfo">
                       <details>
                         <summary>More info</summary>
-                        <p> - by Refsnes Data. All Rights Reserved.</p>
+                        <img src="src/img/wind.png" alt="wind">45
                         <p> - by Refsnes Data. All Rights Reserved.</p>
                         <p> - by Refsnes Data. All Rights Reserved.</p>
                         <p> - by Refsnes Data. All Rights Reserved.</p>
